[web] Reorganize the routes
* The App component defines the routes and makes sure
  that the D-Bus service is working.
* The Main component makes sure that a product is selected.
