+++ web/src/App.jsx
@@ -19,38 +19,84 @@
  * find current contact information at www.suse.com.
  */
 
-import React, { useEffect, useState } from "react";
-import { Outlet, Navigate } from "react-router-dom";
+import React, { useEffect, useReducer } from "react";
 import { useInstallerClient } from "./context/installer";
+import { Outlet, Navigate } from "react-router-dom";
 
+import { PROBING, PROBED, INSTALLING, INSTALLED } from "./client/status";
+
+import DBusError from "./DBusError";
+import ProbingProgress from "./ProbingProgress";
+import InstallationProgress from "./InstallationProgress";
+import InstallationFinished from "./InstallationFinished";
 import LoadingEnvironment from "./LoadingEnvironment";
-import Questions from "./Questions";
 
-import './assets/fonts.scss';
-import "./app.scss";
+const init = status => ({
+  loading: status === null,
+  probing: status === PROBING,
+  probed: status === PROBED,
+  installing: status === INSTALLING,
+  finished: status === INSTALLED,
+  dbusError: null
+});
 
-// TODO: add documentation
+const reducer = (state, action) => {
+  switch (action.type) {
+    case "CHANGE_STATUS": {
+      return { ...state, ...init(action.payload.status) };
+    }
+    case "SET_DBUS_ERROR": {
+      return { ...state, dbusError: action.payload.error };
+    }
+    case "LOAD_PRODUCTS": {
+      return { ...state, products: action.payload.products };
+    }
+    default: {
+      throw new Error(`Unsupported action type: ${action.type}`);
+    }
+  }
+};
+
 function App() {
-  const [product, setProduct] = useState(null);
   const client = useInstallerClient();
+  const [state, dispatch] = useReducer(reducer, null, init);
 
   useEffect(() => {
-    return client.software.getSelectedProduct()
-      .then(setProduct)
-      .catch(() => setProduct(undefined));
+    client.manager.getStatus()
+      .then(status => dispatch({ type: "CHANGE_STATUS", payload: { status } }))
+      .catch(error => dispatch({ type: "SET_DBUS_ERROR", payload: { error } }));
+  }, [client.manager]);
+
+  useEffect(() => {
+    return client.manager.onChange(changes => {
+      console.log("manager has changed", changes);
+
+      if ("Status" in changes) {
+        dispatch({ type: "CHANGE_STATUS", payload: { status: changes.Status } });
+      }
+    });
+  }, [client.manager]);
+
+  useEffect(() => {
+    return client.monitor.onDisconnect(() => {
+      dispatch({ type: "SET_DBUS_ERROR", payload: { error: "Connection lost" } });
+    });
+  }, [client.monitor]);
+
+  useEffect(() => {
+    return client.software.getProducts()
+      .then(products => {
+        dispatch({ type: "LOAD_PRODUCTS", payload: { products } });
+      });
   }, [client.software]);
 
-  // FIXME: improve this
-  if (product === null) return <LoadingEnvironment text="Checking products..." />;
+  if (state.dbusError) return <DBusError />;
+  if (state.loading || !state.products) return <LoadingEnvironment />;
+  if (state.probing) return <ProbingProgress />;
+  if (state.installing) return <InstallationProgress />;
+  if (state.finished) return <InstallationFinished />;
 
-  if (!product) return <Navigate to="products" replace={true} />;
-
-  return (
-    <>
-      <Questions />
-      <Outlet />
-    </>
-  );
+  return <Outlet />;
 }
 
 export default App;
